unpack misc functions
